Refactor DFDA module structure and enhance API client functionality
- Introduced a new `api-client.ts` file to centralize API request handling, including token management and error handling.
- Added `auth.ts` for user authentication and token management, ensuring secure access to DFDA services.
- Created `statistics.ts` to manage markdown statistics retrieval for benefits and problems.
- Implemented `studies.ts` to handle study-related API interactions, including creation and retrieval of studies.
- Developed `treatments.ts` to manage treatment-related functionalities, including fetching conditions and adding treatments.
- Enhanced `variables.ts` to streamline variable management and retrieval processes.
- Updated the main DFDA actions file to export new modules for better organization and maintainability.

This commit improves the modularity and maintainability of the DFDA integration, facilitating easier updates and enhancements in the future.
